disaster-recovery/vaultwarden/pyproject.toml
[project]
name = "vaultwarden"
version = "0.1.0"
description = "A script for backup data and restore"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.17",
    "boto3-stubs~=1.42.17",
    "cryptography>=46.0.3",
    "python-dotenv>=1.2.1",
    "zstandard>=0.25.0",
]
